extract resume_test_route
